Enable arbitrary networks on baremetal nodes
For network isolation testing in OVB, we need to have the ability
to add around 6 nics per baremetal node.  Since it would get
awkward to simply add more parameters for every extra network (and
there are use cases where a different number of networks is needed)
this change replaces the provision and public net parameters with
a couple of new parameters: networks and network_count for the
simple virtual-baremetal and quintupleo cases respectively.

networks should be list of network names to attach to the instances.
It is ignored when using quintupleo.yaml

network_count is the number of networks to attach to the baremetal
instances.  Note that the first two networks will also be attached
to the undercloud instance, so those should be used as the
provisioning and public networks.