Allow use of pre-built bmc images
Previously, the bmc instance was built at deploy time, which has a
couple of potential problems.  First, it adds time to the deploy
process for every single environment, which could be non-trivial
when deploying dozens of environments in a CI situation.  Second,
it introduces additional failure modes if there is a problem
contacting the yum mirrors or pyghmi.  Additionally, using a
pre-built image allows some stabilization of the package versions
involved, rather than always pulling down the latest version of
every dependency.