Remove unnecessary security groups
There were two security groups being created in the Heat templates
that were unnecessary.  all_sg is irrelevant since the ports it
was being applied to must have security groups turned off anyway,
and the private_sg in virtual-baremetal.yaml was no longer actually
applied to any ports.

I left the undercloud security group because it is necessary to
allow access to the undercloud via ssh, and this way the user
doesn't have to configure that manually.  We can revisit this if
it becomes a blocker for things like public cloud deployments.