Reorganize into package and add tox for testing
Moves the functional code into an openstack_virtual_baremetal env
and adds a tox configuration for testing.  Existing unit tests for
deploy.py are moved into the tests subpackage.  Further unit tests
for the other modules will be added in followup commits.

Symlinks from the bin directory are left so the previous workflow
should continue to work as before.