Add support for deploying multiple roles
Allow specifying additional role environment files on the deploy
command-line.  These additional env files will be used to deploy
stacks that are wired into the base stack's resources in such a
way as to allow the instances from all the stacks to be deployed
as a single virtual baremetal environment.

Note that this feature requires use of --quintupleo.