Add env validation to deploy
There's an edge case with role deployments where an env contains
a role, the baremetal_prefix ends in -role, and --id is not used
for deployment.  This will break build-nodes-json due to the
processing it does on baremetal_prefix.  It is also generally a bad
idea anyway because the other roles get built from the base prefix,
so including a role name in the prefix will result in confusing
names like baremetal-control-compute_0.